refactor public api for endpoints, added tests
