+++ bases/api_public/core/fastapi/dependencies/kafka.py
