+++ bases/api_public/core/fastapi/dependencies/session.py
@@ -1,4 +1,4 @@
-from bot_detector.api_public.src.core.config import DB_SEMAPHORE, settings
+from bases.api_public.core.config import DB_SEMAPHORE, settings
 from bot_detector.database import Settings as DBSettings
 from bot_detector.database import get_session_factory
 from sqlalchemy.ext.asyncio import AsyncSession
