+++ bases/api_public/core/fastapi/middleware/metrics.py
