+++ bases/api_public/feedback/schemas.py
@@ -43,3 +43,11 @@ class FeedbackInput(BaseModel):
                 return value
             case _:
                 raise ValueError("Invalid format for player_name")
+
+
+class FeedbackScore(BaseModel):
+    count: int
+    possible_ban: bool
+    confirmed_ban: bool
+    confirmed_player: bool
+    vote: Optional[int] = Field(None, ge=-1, le=1)
