+++ bases/api_public/labels/__init__.py
@@ -1,0 +1,3 @@
+from bases.api_public.labels.routes import router
+
+__all__ = ["router"]
