--- bases/bot_detector/api_public/src/api/__init__.py
@@ -1,6 +1,0 @@
-from fastapi import APIRouter
-
-from . import v2
-
-router = APIRouter()
-router.include_router(v2.router, prefix="/v2")
