--- bases/bot_detector/api_public/src/app/readme.md
@@ -1,6 +1,0 @@
-the model is responsible for all the data handeling
-- getting data from the database
-- handles data logic
-
-the view is responsible for the data representation
-- return format etc
