--- bases/bot_detector/api_public/src/app/views/input/_metadata.py
@@ -1,5 +1,0 @@
-from pydantic import BaseModel
-
-
-class Metadata(BaseModel):
-    version: str
