--- bases/bot_detector/api_public/src/app/views/response/feedback_score.py
@@ -1,8 +1,0 @@
-from pydantic import BaseModel
-
-
-class FeedbackScoreResponse(BaseModel):
-    count: int
-    possible_ban: bool
-    confirmed_ban: bool
-    confirmed_player: bool
