--- bases/bot_detector/api_public/src/app/views/response/report_score.py
@@ -1,9 +1,0 @@
-from pydantic import BaseModel
-
-
-class ReportScoreResponse(BaseModel):
-    count: int
-    possible_ban: bool
-    confirmed_ban: bool
-    confirmed_player: bool
-    manual_detect: bool
