+++ docker-compose.yml
@@ -137,7 +137,7 @@ services:
       context: ./
       dockerfile: ./projects/hiscore_scraper/Dockerfile
       target: production
-    # command: ["uv", "run", "uvicorn", "bot_detector.api_public.src.core.server:app", "--host", "0.0.0.0", "--port", "5000", "--log-level", "warning", "--reload", "--reload-dir", "/app/bot_detector/api_public/"]
+    # command: ["uv", "run", "uvicorn", "bases.api_public.core.server:app", "--host", "0.0.0.0", "--port", "5000", "--log-level", "warning", "--reload", "--reload-dir", "/app/bot_detector/api_public/"]
     networks:
       - botdetector-network
     env_file:
@@ -283,7 +283,7 @@ services:
       target: builder
     # command: ["sleep", "infinity"]
     command: >
-      sh -c "cd ../.. && projects/api_public/.venv/bin/uvicorn bases.bot_detector.api_public.src.core.server:app --host 0.0.0.0 --port 5000 --log-level warning --reload --reload-dir /app"
+      sh -c "cd ../.. && projects/api_public/.venv/bin/uvicorn bases.api_public.core.server:app --host 0.0.0.0 --port 5000 --log-level warning --reload --reload-dir /app"
     env_file:
       - .env
     environment:
