+++ projects/api_public/pyproject.toml
@@ -20,13 +20,13 @@ dependencies = [
 ]
 
 [project.scripts]
-scrape_task_producer = "bot_detector.api_public.core.server:run"
+scrape_task_producer = "bases.api_public.core.server:run"
 
 [tool.hatch.build.hooks.polylith-bricks]
 packages = ["bot_detector"]
 
 [tool.polylith.bricks]
-"../../bases/bot_detector/api_public" = "bot_detector/api_public"
+"../../bases/api_public" = "bot_detector/api_public"
 "../../components/bot_detector/database" = "bot_detector/database"
 "../../components/bot_detector/kafka" = "bot_detector/kafka"
 "../../components/bot_detector/structs" = "bot_detector/structs"
